@@ -8,7 +8,7 @@ description = "Griffe extension for inheriting docstrings."
 authors = [{name = "Timothée Mazzucotelli", email = "dev@pawamoy.fr"}]
 license = {text = "ISC"}
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 keywords = []
 dynamic = ["version"]
 classifiers = [
@@ -17,12 +17,12 @@ classifiers = [
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: 3.13",
+    "Programming Language :: Python :: 3.14",
     "Topic :: Documentation",
     "Topic :: Software Development",
     "Topic :: Utilities",
@@ -55,7 +55,6 @@ source-includes = [
     "scripts",
     "share",
     "tests",
-    "devdeps.txt",
     "duties.py",
     "mkdocs.yml",
     "*.md",
@@ -65,4 +64,41 @@ source-includes = [
 [tool.pdm.build.wheel-data]
 data = [
     {path = "share/**/*", relative-to = "."},
+]
+
+[tool.uv]
+dev-dependencies = [
+    # dev
+    "editables>=0.5",
+
+    # maintenance
+    "build>=1.2",
+    "git-changelog>=2.5",
+    "twine>=5.1",
+
+    # ci
+    "duty>=1.4",
+    "ruff>=0.4",
+    "pytest>=8.2",
+    "pytest-cov>=5.0",
+    "pytest-randomly>=3.15",
+    "pytest-xdist>=3.6",
+    "mypy>=1.10",
+    "types-markdown>=3.6",
+    "types-pyyaml>=6.0",
+
+    # docs
+    "black>=24.4",
+    "markdown-callouts>=0.4",
+    "markdown-exec>=1.8",
+    "mkdocs>=1.6",
+    "mkdocs-coverage>=1.0",
+    "mkdocs-gen-files>=0.5",
+    "mkdocs-git-revision-date-localized-plugin>=1.2",
+    "mkdocs-literate-nav>=0.6",
+    "mkdocs-material>=9.5",
+    "mkdocs-minify-plugin>=0.8",
+    "mkdocstrings[python]>=0.25",
+    # YORE: EOL 3.10: Remove line.
+    "tomli>=2.0; python_version < '3.11'",
 ]
